requirements.txt
streamlit
pillow=10.3.0
tensorflow==2.16.1
matplotlib
numpy
joblib==1.4.0
io
sys
